Change how expressions are represented.
Instead of having different types for each kind of expression, with the
implicit requirement that they must be castable to the generic expr type, we
now have a single union for expr type.

This incurs a few more bytes overhead for some expressions, but avoids having
to keep types in sync, avoids having to cast between different types, and
makes it explicit what all the possible expression types are.