Switch back to exposing the fields of value.
Instead of using setter and getter functions. This time, use a single
Value type for both structs and unions, and don't try to pack things in a
complex way. Adding some overhead is fine, because we don't expect this
interface to be used in performance critical scenarios and if we need to, we
can convert to more efficient representations of values internally without
changing the interface.