Switch to new approach for storing symbol info.
Rather than using a location id for every token in the program, use location
ids just for actions and expressions and structured information about the
tokens.

This adds an id field to FblcActn and FblcExpr to make it easy to look up
symbol information for actions and expressions. If the extra space is a
problem, the id field could be merged with, for example, the tag fields.

The code is much cleaner this way.