Changed syntax for struct and union values.
Struct: type@(args)
Union: type@(field: arg)

Now we can distinguish syntactically between struct_value and apply, so we
know how far we have to evaluate the function when type checking. And now the
syntax of field choice is more consistent, with the colon always after the
field and the pair inside parens.