Retain instructions holder on stack frame, not data stack.
Often the instructions we execute are owned by a function or process value. We
need to retain that function or process value for the duration of execution of
those instructions so they aren't freed out from under us.

Retain the function/process value as part of the stack frame rather than on
the data stack. This way we no longer need the awkward RELEASE instruction.