Change type of PUT process to return *()!.
It was always a little awkward to return the input value. Now that types are
values instead of declared things, it's much more natural to have a put
process return the unit type. And we already do this kind of thing in the
compiler with literals, so it's not setting a new precedent.