Remove fble-Snake.c, fble-snake.c, and fble-snake.fble.
fble-Snake.c is obsoleted by fble-app.c. It doesn't seem worth the effort to
maintain an ncurses version of fble-app.c given we expect the App interface to
change quit a bit. If we eventually want to support an ncurses version of
fble-app.c, for easier use with linux consoles and ssh for example, we can
always add it back later.

It also doesn't seem worth keeping around fble-snake.fble.