Change App% drawing API to draw from scratch each frame.
Instead of drawing on top of the previous frame. It doesn't work to draw on
top of the previous frame if we are double buffering and trying to draw on the
back buffer while the front buffer is used for display.

Update snake to redraw itself entirely every frame instead of doing
incremental updates to the screen.