Profile based on random sampling.
So we don't have to keep track of ticks since last profile, and as way of
excuse for inlining FbleProfileTick. I think this is much faster (on the order
of 30-60s saved on 4 minutes of fble bench), and it should give better
profiling results because we don't have to worry about funny biasing.

Note that the 'profile time' is now in units of thousands of instructions
executed.