Separate FbleForkProfileThread from FbleNewProfileThread.
It's silly to have one function that has two mutually exclusive ways to
specify the arguments that results in two almost entirely mutually exclusive
implementations, while requiring the caller to carry both a profile thread and
a profile.

Split it into two separate functions, which makes a lot more sense.