Reorganize implementation of func apply instruction.
In such a way that the value used to retain the scope also keeps alive all of
the arguments. In preparation for using the native stack for eval.

Performance got a bunch worse, but I expect it to get more better soon, so I
think that's okay.

fble bench: 3m22s -> 3m59s.