Fix bug in FbleApply.
We were allocating an FbleInstrBlock on the stack in FbleApply. Even though
that code block isn't retained by the result, there are references to it from
the thunk after FbleApply returns if gc hasn't collected the thunk yet. That's
a dangling pointer that leads to a segfault when the thunk is eventually
collected.

Rewrite FbleApply to avoid allocating any instruction block at all. It seems
just as easy to do the application directly rather than through instructions
now.