Revert back to non-native stacks implementation.
Use of native stacks is not practical because of the requirement of allocating
a large enough stack for every thread. Go back to managing stacks explicitly
so we can dynamically grow them as appropriate.

If we are going to manage stacks explicitly, no need to use pthreads anymore
for scheduling.