Distinguish between types (a, b){c;} and (a)(b){c;}.
As an experiment to see how much we rely on partial or over application of
functions. With the idea that partial applications have real performance cost,
and it's worth making the programmer aware of that.

This changes the compiler only. We haven't yet updated the instructions to
support multi arg application, so everything you try to run will crash at this
point.