Pull compile phase out of FbleEval.
I think it's better for FbleCompile to be separate from FbleEval. It lets us
re-use a compiled program if desired, means you can compile a program without
having to provide a value heap, and makes the flow better match what I expect
we will have if we start compiling with llvm.