Switch Snake to use S@ for the snake instead of P@.
P@ is a more precise representation, because we know we always have at least
one element and often want to look at the head, but S@ is much more prevelant
and has better library support. I'm thinking that better library support
outweighs the slight loss in precision of the type being used.