Remove FbleLoc from FbleTc.
Some of the locations are specific to the particular kind of FbleTc. For
example, For FbleAccessTc we want the location to refer to the '.', not the
start of the expression. For this reason, it seems cleaner to give every Tc
that needs a location an explicit custom location rather than a generic
location that we have to preserve everywhere but for the most part has limited
applicability.