Fix issue where FullGc wasn't doing full GC.
FullGc was only removing references unreachable at the time you start the
FullGC. But it's possible that the on_free for an object releases other
references. Change FullGc to continue doing GC for as long as any object
is freed.

This exposes an issue with the memory growth tests that needs to be
investigated.