Switch to a different representation for FBLE_UNION_SELECT_TC.
Instead of having two lists, one for choices and one for branches, have a
single list for choices where some of the choices can point to the same value.
I think this is a little more straight forward of a representation, and it
doesn't seem to make the code any messier.

This way we can represent union select with a single list, which makes it more
natural to allocate FBLE_UNION_SELECT_VALUE in a single block of memory once
we have it take the place of FBLE_UNION_SELECT_TC.