Change Stack to FbleThunkValueTc.
There is a bunch of work needed updating tracking of references that hasn't
been done yet. But this seems like a decent place to check something in given
that a bunch of tests are passing and I think all that's left that's needed is
to fix the reference tracking issues.