Remove del_ref from heap API.
I think it's a worth while API to have, but currently its a no-op in our only
implementation of the heap API. That means we have no way test that we are
using it correctly. Rather than waste time maintaining code to call del_ref
that is probably broken regardless, remove the API entirely. We can add it
back at the point when we want it again.