Clean up computation from completed thunks.
Otherwise we'll end up holding on to a lot of memory unnecessarily whenever
there is a thunk value.

With this change, all tests and benchmarks are now passing again.

fble bench: 4m32.575s

Profiling suggests the increase in benchmark time is due primarily to the
extra time spent doing GC given that the number of GC managed allocations has
doubled from allocating the stack as GC objects, and the cost of calling
FbleValueAddRef a lot more often.