Generate prototypes as part of CompileEnv.
Otherwise we weren't outputting the functions in proper order.

Also: temporarily remove some logic from Md5 to get a sense of how much memory
will be required for full Md5. Answer is a lot. But good news is that it takes
less than .02 seconds to execute the compiled C code compared to 25 seconds
interpreted in fble. That's 1000x speedup.