WIP: Avoid calling module bodies more than once.
By transforming the result of a module for each instance instead of rerunning
the body of the module for each instance. Because we the result is the size of
the module interface and the body is the size of the module implementation,
and we expect the interface to be much smaller than the implementation.

But we still have some memory bug that needs to be tracked down.