Remove notion of TIME_SLICE for yielding.
I think it's probably a good idea to have a notion of a time slice to ensure
fairness among threads, but we don't currently have a decent way to test it,
and the spec isn't entirely clear. I'd rather wait until we run into this in
practice to figure out a proper solution.