Remove notion of private/public access control for modules.
I like the theory behind public and private modules, but I'm not convinced the
use of name Foo*.tcl is the way I want to go. Rather than have to do extra
work to support the currently unused framework for private/public access for
compilation, remove the private/public stuff from the language spec for now.
We can always add it back in later when we start to want to use it more, and
after having a chance to rethink the details.