Factor out 'fbleobj' build rule.
To make it easier to more comprehensively switch between compilation via
c and compilation via assembly.

Change fble-compile to take [--export ...] as first command line
argument instead of intermixed with other arguments. It turns out this
change has nothing really to do with anything, but it's done and I
think a little cleaner, so may as well keep it.