Change syntax for stmt based union select default.
Instead of having an 'if' statement, treat it as a variation on a union
select expression with a different way to specify the default value.

This way we can use a syntax consistent with the rest of the syntax in
fble and allow multiple non-default branches in the union select instead
of a single non-default branch.