Fix pathological memory issue with profiling.
Remove the hash table optimization from the profiling code, which was
leading to absurd memory use in the fork-race pathalogical case.

This slows profiling down a fair amount (from 6m11s to 8m33s on fble
bench compiled). We can come up with a more memory efficient
optimization in the future if desired.