Clean up coordinate systems and signs.
Switch to a left handed coordinate system, which feels more natural for
me for the camera: x and y are in the expected direction the camera
point in the +z direction.

Explain all the choices of signs. This fixes the failing rotation test
and works as expected now in the graphics app.