Add tracking of self time in the profiler.
We sort of just dump the self time as an extra column in the profiling
report for the time being. Eventually we ought to clean that up so it's
sorted, or so we don't include a bunch of 0 columns, or whatever makes
sense.

Optimize fble-bench uses of Int|xxx that could be factored out.

fble bench interpreted: 7m43s -> 6m46s
fble bench compiled:    4m57s -> 4m21s