Reorganize fble-deps implementation.
Track dependencies in FbleLoad instead of reconstructing them from the
loaded program. This way we don't have to duplicate the search logic in
case of multiple entries in the search path, we can add dependency
generation as part of other tools such as fble-compile instead of
requiring a separate pass, and we can improve the list of dependencies
later on to include more details about directories searched.