Move subdirectories out of fble/ directory.
Because we want include files to live under an include/fble/* directory,
and it seems silly to have fble/include/fble/*, remove the leading fble/
directory entirely. That directory doesn't add much, especially now that
fble is the only language of fable I expect to be working on for the
near future.

Also: fix a broken reference to sat-tests in pkgs/sat/build.tcl.