Use managed stack only for tail calls.
We can use the native stack for normal function calls, because the
caller is holding on to the function and args for us. Save a stack
allocation and extra retain/release for func and args in the normal call
case by not putting them on the managed stack.

On my pi:

Intepreted, compared to previous measurement:
  /Md5/Bench%: 1m13s ==> 1m06s
  /Sat/Bench%: 1m48s ==> 1m44s

aarch64 compiled:
  md5-bench: 44s
  sat-bench: 1m13s