Change FbleReleaseInstr to take multiple values.
It's pretty common to have back-to-back release instructions. Change the
FbleReleaseInstr to take a vector of values to release instead of a
single value.

The main motivation is so the fble-disassemble code looks nicer for the
CompiledCode tutorial, but maybe it will help with performance and
compactness of generated code too.