Turn off SGR for usage text generation.
So that you can pipe the output of --help to less and it shows up
reasonable.

Looks like you get bold formatting with less this way, just not with
plain text. The other option would be to require use of less -r, but
that seems non-standard. For nice formatting, users can use man pages
instead of --help.