Remove timeout and fble-sat-bench.
Having a timeout based on number of assignments tried is not useful in
practice. There is no way to predict how much wall clock time a
particular timeout is associated with.

Setting a benchmark based on timeout is not useful for benchmarking the
sat solver, because changes to the sat solver algorithm change the
interpretation of timeout.

To simplify things, remove timeout and fble-sat-bench entirely. We can
do sat benchmarking on solvable problems from sat benchmark sets
instead, using the timeout command to give meaningful timeouts.