Introduce an Fbld@ monad.
Use a concrete Fbld@ monad for processing Fbld documents. The result of
processing is Text@ instead of abstract side effects.

It's a bit more convenient to work with concrete monads than abstract
monads. This paves the way for defining fbld commands that can add new
commands to the environment on the fly, so we can pull front ends and
back ends out of core fbld implementation entirely.