Remove redundant calls to FbleStrictValue.
Change FbleStructValueField, FbleUnionValueArg, FbleUnionValueTag, and
FbleUnionValueField to return a special value if the argument is
undefined. Use this instead of an explicit call to FbleStrictValue to
detect the undefined value case.

Don't allow NULL struct or union fields anymore.

Time to generate fble.html with fbld improved by 8% from 96s to 89s.