Avoid capturing type variables in scope.
We can always compute type values on the fly. Avoid capturing them in
function scopes.

In preparation for merging multi-arg function values at compile time.

In preparation for avoiding the need to define a type var value for poly
values at runtime so poly function values can be merged with multi-arg
functions.

And to save a little space at runtime for function values.