Be slightly more lenient on memory test.
For whatever reason, on windows we sometimes are 4KB off after getting
max memory usage. It's so much smaller than we are off when we have a
real memory leak though, so it should be fine.

With this we are now able to compile on windows out of the box.