Check for required tail call space on function allocation.
Instead of when using the tail call buffer. Because we expect to
allocate functions a lot less often than when we call functions.

This is in preparation for being able to inline the FbleTailCall logic
into backends. It will be much simpler if the backends don't have to
worry about checking or resizing the tail call buffer.