Avoid hardcoding the build path to profile.c in output profiles.
The build path changes depending on the build directory in use.
Hardcoding the build directory into output profiles makes it harder to
test, because the expected profile changes depending on the build
directory.

I doubt putting 'profile.c' for the source file for [root] profile
block is useful for anyone anyway.