Implement /Core/String/Eq% functions.
As specializations of /Std/List/Eq% functions.

Notably: Change Split function to be 'Eq' based split instead of
'Search' based split. In theory people can use the list implementation
of search based split for strings just as easily. It's convenient to use
the eq based split for the few cases we are using it in practice.