Remove \r and \t support from StrE.
Having \n makes sense because newline is part of the convention for line
breaks adopted by the fble standard library. \r and \t on the other hand
feel a bit too arbitrary to assign them, and not a bunch of other
potential characters in non-ascii character sets, as escape sequences.