Change definition of Show@ interface.
To allow for more flexibility in the output stream type, so we can avoid
the quadratic blowups you can get when outputting to String@ and we can
show things to output streams incrementally.

This just changes the definition of Show@, not any of the users of it.
Everything will be broken now and for a while as we switch users over to
the new definition.